Splitting selectedItemOptions into components
