+++ src/store/fileBrowserStore.ts
@@ -75,7 +75,7 @@ export const chatFiles = ref<File[]>();
 export const allSharedContent = ref<SharedFileInterface[]>([]);
 export const chatsWithFiles = ref<Chat[]>();
 
-export const fileBrowserTypeView = ref<ViewType>('LIST');
+export const fileBrowserTypeView = ref<ViewType>(ViewType.LIST);
 export const accessDenied = ref(false);
 export const chatFilesBreadcrumbs = ref([])
